PDCL-7517 - move createDataCollectionRequestPayload from core to utils/request
