+++ src/core/index.js
@@ -35,8 +35,10 @@ import injectSendBeaconRequest from "./network/requestMethods/injectSendBeaconRe
 import createLogger from "./createLogger";
 import createEventManager from "./createEventManager";
 import createCookieTransfer from "./createCookieTransfer";
-import createDataCollectionRequestPayload from "./edgeNetwork/requests/createDataCollectionRequestPayload";
-import createDataCollectionRequest from "../utils/request";
+import {
+  createDataCollectionRequest,
+  createDataCollectionRequestPayload
+} from "../utils/request";
 import injectSendEdgeNetworkRequest from "./edgeNetwork/injectSendEdgeNetworkRequest";
 import injectProcessWarningsAndErrors from "./edgeNetwork/injectProcessWarningsAndErrors";
 import isRequestRetryable from "./network/isRequestRetryable";
