+++ src/utils/request/createDataCollectionRequestPayload.js
@@ -10,8 +10,8 @@ OF ANY KIND, either express or implied. See the License for the specific languag
 governing permissions and limitations under the License.
 */
 
-import { createRequestPayload } from "../../../utils/request";
-import { createAddIdentity } from "../../../utils/request/contentModifiers";
+import createRequestPayload from "./createRequestPayload";
+import { createAddIdentity } from "./contentModifiers";
 
 export default () => {
   const content = {};
