+++ src/utils/request/index.js
@@ -11,5 +11,6 @@ governing permissions and limitations under the License.
 */
 
 export { default as createDataCollectionRequest } from "./createDataCollectionRequest";
+export { default as createDataCollectionRequestPayload } from "./createDataCollectionRequestPayload";
 export { default as createRequest } from "./createRequest";
 export { default as createRequestPayload } from "./createRequestPayload";
